+++ tests/Sabre/DAV/TestPlugin.php
@@ -23,7 +23,7 @@ class TestPlugin extends ServerPlugin {
 
     function initialize(Server $server) {
 
-        $server->on('beforeMethod', [$this, 'beforeMethod']);
+        $server->on('beforeMethod:*', [$this, 'beforeMethod']);
 
     }